[Fix] AuthFacade 도입
