+++ src/main/java/troublog/backend/domain/auth/service/AuthService.java
@@ -32,76 +32,9 @@ import java.time.Duration;
 @RequiredArgsConstructor
 public class AuthService {
 
-	private final AuthenticationManager authenticationManager;
-	private final UserQueryService userQueryService;
-	private final UserCommandService userCommandService;
-	private final PasswordEncoder passwordEncoder;
 	private final JwtProvider jwtProvider;
 
-	@Value("${spring.profiles.active}")
-	private String profilesActive;
-
-	@Transactional
-	public Long register(RegisterDto registerDto, HttpServletRequest request) {
-
-		String clientEnvType = request.getHeader("EnvType");
-
-		// 프론트 환경변수 체크
-		jwtProvider.checkEnvType(clientEnvType);
-
-		// 닉네임 중복 체크
-		boolean isDuplicatedNickname = userQueryService.existsByNickname(registerDto.nickname());
-		if (isDuplicatedNickname) {
-			throw new UserException(ErrorCode.DUPLICATED_NICKNAME);
-		}
-
-		// 비밀번호 인코딩
-		String encodedPassword = passwordEncoder.encode(registerDto.password());
-
-		User user = UserConverter.toEntity(registerDto, encodedPassword);
-
-		return userCommandService.save(user);
-	}
-
-	@Transactional
-	public LoginResDto login(LoginReqDto loginReqDto, HttpServletRequest request, HttpServletResponse response) {
-
-		String clientEnvType = request.getHeader("EnvType");
-
-		// 프론트 환경변수 체크
-		jwtProvider.checkEnvType(clientEnvType);
-
-		// 유저 확인
-		User user = userQueryService.findUserByEmailAndIsDeletedFalse(loginReqDto.email());
-
-		// 비밀번호 검증
-		if (!passwordEncoder.matches(loginReqDto.password(), user.getPassword())) {
-			throw new UserException(ErrorCode.INVALID_USER);
-		}
-
-		// Authentication 객체 생성
-		// 유저 이메일(아이디), 비밀번호 외에 유저 아이디, 닉네임, 프론트 쪽 환경변수도 claim 으로 넣어주는 CustomAuthenticationToken
-		CustomAuthenticationToken authenticationToken =
-			CustomAuthenticationToken.unauthenticated(loginReqDto.email(), loginReqDto.password(), user.getId(),
-				clientEnvType,
-				user.getNickname());
-
-		Authentication authentication = authenticationManager.authenticate(authenticationToken);
-
-		// 액세스토큰, 리프레시토큰 생성
-		String accessToken = jwtProvider.createAuthToken(authentication);
-		String localToken = jwtProvider.createAuthToken(authentication);
-		String refreshToken = jwtProvider.createRefreshToken(authenticationToken);
-
-		// 리프레시 토큰 Set-Cookie로 내려줌
-		setCookieRefreshToken(refreshToken, response);
-
-		return (profilesActive.equals(EnvType.LOCAL.getEnvType()) || clientEnvType.equals(EnvType.LOCAL.getEnvType()))
-			? LoginResDto.localReturn(user.getId(), accessToken, refreshToken, localToken)
-			: LoginResDto.nonLocalReturn(user.getId(), accessToken, refreshToken);
-	}
-
-	private void setCookieRefreshToken(String refreshToken, HttpServletResponse response) {
+	public void setCookieRefreshToken(String refreshToken, HttpServletResponse response) {
 		ResponseCookie cookie = ResponseCookie.from("refreshToken", refreshToken)
 			.httpOnly(true)
 			.secure(false)
@@ -113,34 +46,6 @@ public class AuthService {
 		response.setHeader(HttpHeaders.SET_COOKIE, cookie.toString());
 	}
 
-	@Transactional
-	public String reissueAccessToken(HttpServletRequest request) {
-
-		String clientEnvType = request.getHeader("EnvType");
-
-		// 프론트 환경변수 체크
-		jwtProvider.checkEnvType(clientEnvType);
-
-		// jwtProvider 에서 액세스토큰의 만료시간, 리프레시토큰의 유효성 검증
-		Long userId = jwtProvider.reissueAccessToken(request);
-
-		// 새로운 액세스토큰에 넣어 줄 유저 정보 조회
-		User user = userQueryService.findUserByIdAndIsDeletedFalse(userId);
-
-		//새로운 CustomAuthenticationToken 객체 생성
-		CustomAuthenticationToken authenticationToken =
-			CustomAuthenticationToken.unauthenticated(
-				user.getEmail(),
-				user.getPassword(),
-				user.getId(),
-				clientEnvType,
-				user.getNickname());
-
-		Authentication authentication = authenticationManager.authenticate(authenticationToken);
-
-		// 새로운 액세스토큰 생성
-		return jwtProvider.createAuthToken(authentication);
-	}
 
 	@Transactional
 	public void logout(HttpServletRequest request, HttpServletResponse response) {
@@ -166,19 +71,5 @@ public class AuthService {
 			.build();
 
 		response.setHeader(HttpHeaders.SET_COOKIE, deleteCookie.toString());
-	}
-
-	@Transactional(readOnly = true)
-	public void checkDuplicateEmail(String email, HttpServletRequest request) {
-
-		String clientEnvType = request.getHeader("EnvType");
-
-		// 프론트 환경변수 체크
-		jwtProvider.checkEnvType(clientEnvType);
-
-		boolean isDuplicated = userQueryService.existsByEmail(email);
-		if (isDuplicated) {
-			throw new UserException(ErrorCode.DUPLICATED_EMAIL);
-		}
 	}
 }
